Change Float fields to Double for performance and precision reasons

+++ app/src/main/java/pl/bk20/forest/domain/model/DayParameters.kt
@@ -14,5 +14,5 @@ data class DayParameters(
 
     val stepLength: Int,
 
-    val pace: Float
+    val pace: Double
 )
